requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
app-commands==0.1.0
async==0.6.2
asyncio==3.4.3
attrs==25.3.0
click==8.2.1
DateTime==5.5
discord==2.3.2
discord.py==2.5.2
frozenlist==1.6.0
idna==3.10
logging==0.4.9.6
multidict==6.4.4
propcache==0.3.1
python-dotenv==1.1.0
pytz==2025.2
setuptools==80.8.0
timedelta==2020.12.3
yarl==1.20.0
zope.interface==7.2
